separate views and models
