+++ flask/project/__init__.py
@@ -1,5 +1,4 @@
 from flask import Flask, render_template
-from flask_sqlalchemy import SQLAlchemy
 
 def create_app():
 	app = Flask(__name__)
@@ -7,15 +6,7 @@ def create_app():
 	return app
 
 app = create_app()
-db = SQLAlchemy(app)
 
 from .models import *
 
-db.create_all()
-db.session.commit()
-
-@app.route('/')
-def index():
-	db.session.add(Numbercol(1))
-	db.session.commit()
-	return render_template('index.html')
+from .views import *
