+++ flask/project/models.py
@@ -1,8 +1,16 @@
-from project import db
+from project import app
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy(app)
+
 
 class Numbercol(db.Model):
-	uid = db.Column(db.Integer, primary_key=True)
-	numbercol = db.Column(db.Integer)
+    uid = db.Column(db.Integer, primary_key=True)
+    numbercol = db.Column(db.Integer)
 
-	def __init__(self, numbercol):
-		self.numbercol = numbercol
+    def __init__(self, numbercol):
+        self.numbercol = numbercol
+
+
+db.create_all()
+db.session.commit()
