+++ flask/project/static/css/style.css
@@ -1,0 +1,3 @@
+body {
+    margin: 10px;
+}
