+++ flask/project/views.py
@@ -1,0 +1,6 @@
+from flask import render_template
+from project import app
+
+@app.route('/')
+def index():
+	return render_template('index.html')
